refactor: Point 캐싱
